fix : server 타입에 맞게 타입 수정 및 공통함수 분리
